feat: use CBOR instead
Change the code with minimal changes to do CBOR encoding and decoding intead
of simple-dag. This should show that a CBOR parser that only supports DAG-CBOR
can also be implemented with very little code.

It is without encoding/decoding CIDs as this also doesn't work on master and
I didn't want to spend time fixing on it as this is really just a proof of
concept.